feat(TDP-4285): First OS integ test based on api * Move DataPrepAPIHelper into helper package * Correct dataprep-api dependency in dataprep-test-api
