+++ dataprep-api/src/main/java/org/talend/dataprep/helper/DataPrepAPIHelper.java
@@ -1,21 +1,16 @@
-package org.talend.dataprep;
+package org.talend.dataprep.helper;
 
 import io.restassured.RestAssured;
 import io.restassured.http.Header;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
 import org.apache.commons.io.IOUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.ConfigurableEnvironment;
 import org.springframework.stereotype.Component;
 
 import java.nio.charset.Charset;
 
 @Component
 public class DataPrepAPIHelper {
-
-    @Autowired
-    protected ConfigurableEnvironment environment;
 
     private static RequestSpecification given() {
         return RestAssured.given().log().all(true);
@@ -29,13 +24,13 @@ public class DataPrepAPIHelper {
      * @return the response
      * @throws java.io.IOException if creation isn't possible
      */
-    public Response uploadDataset(String filename, String datasetName) throws java.io.IOException {
+    public Response uploadDataset(String filename, String datasetName, String envURL) throws java.io.IOException {
         Response response =
                 given().header(new Header("Content-Type", "text/plain"))
                         // FIXME : this way of sending datasets through Strings limits the dataset size to the JVM available memory
                         .body(IOUtils.toString(DataPrepAPIHelper.class.getResourceAsStream(filename), Charset.defaultCharset()))
                         .when()
-                        .post(environment.getProperty("run.environment.url") + "/api/datasets?name=" + datasetName);
+                        .post(envURL + "/api/datasets?name=" + datasetName);
         return response;
     }
 
@@ -45,10 +40,10 @@ public class DataPrepAPIHelper {
      * @param dataSetId the dataset to delete.
      * @return the response
      */
-    public Response deleteDataSet(String dataSetId) {
+    public Response deleteDataSet(String dataSetId, String envURL) {
         return given().
                 when().
-                delete(environment.getProperty("run.environment.url") + "/api/datasets/" + dataSetId);
+                delete(envURL + "/api/datasets/" + dataSetId);
     }
 
     /**
@@ -56,7 +51,7 @@ public class DataPrepAPIHelper {
      *
      * @return the response.
      */
-    public Response getDatasetList() {
-        return given().get(environment.getProperty("run.environment.url") + "/api/datasets/");
+    public Response getDatasetList(String envURL) {
+        return given().get(envURL + "/api/datasets/");
     }
 }
