+++ dataprep-test-api/pom.xml
@@ -37,6 +37,7 @@
             <artifactId>dataprep-api</artifactId>
             <groupId>org.talend.dataprep</groupId>
             <version>2.2.0-SNAPSHOT</version>
+            <classifier>lib</classifier>
         </dependency>
         <!-- logging stuff -->
         <dependency>
